deleted all the conditional classes from the style
The majority of the file - something like 100kb - was just an enumeration of various colors and margins combined with different screen sizes. I feel like that's not really what css is meant for. I dunno. Weird. Why minify the css and remove all the spacing but then fill it up with so much redundancy?