Conditional gain re-establishment.
Without this line, it's clicky. (Still don't know why.)
With this line, it's super duper clicky with early release of high attacktime settings.

As a compromise, this line only triggers when the attack time is high enough that the major early-release clickiness would be more annoying than the minor constant clickiness.

I imagine that it would be uncommon to hold the button for significantly less than 0.1 seconds.